requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image Processing
opencv-python-headless>=4.5.0
Pillow>=8.0.0

# Data Handling
requests>=2.25.0

# Visualization & Web App
streamlit>=1.28.0
plotly>=5.0.0
folium>=0.12.0
streamlit-folium>=0.13.0

# Geospatial & Weather
geopandas>=0.10.0
rasterio>=1.3.0
xarray>=0.20.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
python-dotenv>=0.19.0

# Optional for deployment
gunicorn>=20.1.0 